requirements.txt
ib-insync==0.9.64
matplotlib==3.2.1
numpy==1.19.4
pandas==1.0.4
psycopg2-binary==2.8.5
